+++ quantlib/quotes.pyx
@@ -43,5 +43,5 @@ cdef class SimpleQuote(Quote):
             else:
                 return None
 
-        def __set__(self, float value):
+        def __set__(self, double value):
             (<_qt.SimpleQuote*>self._thisptr.get()).setValue(value)
