+++ quantlib/termstructures/volatility/equityfx/black_vol_term_structure.pyx
@@ -26,7 +26,7 @@ cdef class BlackConstantVol(BlackVolTermStructure):
     def __init__(self,
         Date reference_date,
         Calendar calendar,
-        float volatility,
+        double volatility,
         DayCounter daycounter
     ):
